changes in removal handling

+++ MyHoard/MyHoard/Services/ItemService.cs
@@ -32,15 +32,27 @@ namespace MyHoard.Services
             }
         }
 
-        public void DeleteItem(Item item)
+        public void DeleteItem(Item item, bool forceDelete = false)
         {
+
+            if(forceDelete || (String.IsNullOrEmpty(item.PythonId)
+                    && String.IsNullOrEmpty(item.Java1Id) && String.IsNullOrEmpty(item.Java2Id)))
+            {
+                databaseService.Delete(item);
+            }
+            else
+            {
+                item.ToDelete = true;
+                ModifyItem(item);
+            }
+
             MediaService ms = IoC.Get<MediaService>();
             foreach(Media m in ms.MediaList(item.Id, false, false))
             {
-                m.ToDeleteFromIS = true;
+                m.ToDelete = true;
                 ms.ModifyMedia(m);
             }
-            databaseService.Delete(item);
+            
         }
 
         public Item ModifyItem(Item item)
@@ -69,9 +81,12 @@ namespace MyHoard.Services
             return databaseService.ListAll<Item>();
         }
 
-        public List<Item> ItemList(int collectionId)
+        public List<Item> ItemList(int collectionId, bool withDeleted=false)
         {
-            return databaseService.ListAllTable<Item>().Where(i => i.CollectionId == collectionId).ToList();
+            if (withDeleted)
+                return databaseService.ListAllTable<Item>().Where(i => i.CollectionId == collectionId).ToList();
+            else
+                return databaseService.ListAllTable<Item>().Where(i => (i.CollectionId == collectionId && i.ToDelete == false)).ToList();
         }
 
     }
